+++ Cargo.toml
@@ -7,7 +7,7 @@ members = [
 	"libs/core/libs/shared",
 	"libs/core/libs/test_utils",
 	"libs/c_interface_v2",
-	"libs/core_external_interface",
+	"libs/lb_external_interface",
 	"libs/editor/egui_editor",
 	"server/server",
 	"clients/cli",
