+++ clients/android/app/src/test/java/app/lockbook/CancelSubscriptionTest.kt
@@ -13,7 +13,7 @@ class CancelSubscriptionTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
