+++ clients/android/app/src/test/java/app/lockbook/DeleteFileTest.kt
@@ -14,7 +14,7 @@ class DeleteFileTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
