+++ clients/android/app/src/test/java/app/lockbook/ExportDrawingToDiskTest.kt
@@ -15,7 +15,7 @@ class ExportDrawingToDiskTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
