+++ clients/android/app/src/test/java/app/lockbook/GetRootTest.kt
@@ -12,7 +12,7 @@ class GetRootTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
