+++ clients/android/app/src/test/java/app/lockbook/GetUsageTest.kt
@@ -11,7 +11,7 @@ class GetUsageTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
