+++ clients/android/app/src/test/java/app/lockbook/SyncAllTest.kt
@@ -13,7 +13,7 @@ class SyncAllTest {
         @BeforeClass
         @JvmStatic
         fun loadLib() {
-            System.loadLibrary("lockbook_core_external_interface")
+            System.loadLibrary("lb_external_interface")
         }
     }
