--- clients/android/core/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="app.lockbook.core" />
