+++ clients/android/lb-rs/build.gradle
@@ -1,0 +1,52 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: "org.jmailen.kotlinter"
+
+android {
+    compileSdkVersion 32
+
+    defaultConfig {
+        minSdkVersion 22
+        targetSdkVersion 32
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles 'consumer-rules.pro'
+    }
+
+    lintOptions {
+        warningsAsErrors true
+        htmlReport false
+        xmlReport false
+        abortOnError true
+        lintConfig file('../lint.xml')
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    kotlinter {
+        disabledRules = ["no-wildcard-imports", "parameter-list-wrapping"]
+    }
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    // Support Libraries
+    implementation "androidx.appcompat:appcompat:$appcompat_version"
+
+    // Android KTX
+    implementation "androidx.core:core-ktx:$core_ktx_version"
+
+    // Testing Libraries
+    testImplementation "junit:junit:$junit_version"
+    androidTestImplementation "androidx.test.ext:junit:$junit_test_version"
+    androidTestImplementation "androidx.test.espresso:espresso-core:$espresso_test_version"
+}
+repositories {
+    mavenCentral()
+}
