+++ clients/android/lb-rs/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="app.lockbook.core" />
