+++ clients/android/settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name='Lockbook'
 include ':app'
-include ':core'
+include ':lb-rs'
