+++ libs/lb_external_interface/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "lb_external_interface"
+version = "0.8.2"
+edition = "2021"
+
+[lib]
+name = "lb_external_interface"
+crate-type = ["lib", "staticlib", "cdylib"]
+
+[dependencies]
+basic-human-duration = "0.2.0"
+time = "0.3.20"
+crossbeam = "0.8.1"
+jni = { version = "0.13.1", default-features = false }
+lazy_static = "1.4.0"
+lockbook-core = { path = "../core" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.44"
+strum = "0.19.5"
+strum_macros = "0.19.4"
+tracing = "0.1.5"
