+++ utils/dev-tool/src/apple.rs
@@ -5,7 +5,7 @@ use std::io::Write;
 use std::process::{Command, Stdio};
 
 const LIB_NAME_HEADER: &str = "lockbook_core.h";
-const LIB_NAME: &str = "liblockbook_core_external_interface.a";
+const LIB_NAME: &str = "liblb_external_interface.a";
 
 pub fn run_swift_tests(tool_env: &ToolEnvironment) {
     dotenvy::from_path(utils::local_env_path(&tool_env.root_dir)).unwrap();
@@ -26,7 +26,7 @@ pub fn run_swift_tests(tool_env: &ToolEnvironment) {
 }
 
 pub fn make_swift_test_lib(tool_env: &ToolEnvironment) {
-    let ext_iface_dir = utils::core_external_interface_dir(&tool_env.root_dir);
+    let ext_iface_dir = utils::lb_external_interface_dir(&tool_env.root_dir);
     let c_interface_file = ext_iface_dir
         .join("src/c_interface.rs")
         .to_str()
