+++ utils/dev-tool/src/utils.rs
@@ -47,8 +47,8 @@ pub fn swift_lib<P: AsRef<Path>>(root: P) -> PathBuf {
         .join("clients/apple/CLockbookCore/Sources/CLockbookCore/lib")
 }
 
-pub fn core_external_interface_dir<P: AsRef<Path>>(root: P) -> PathBuf {
-    root.as_ref().join("libs/core_external_interface")
+pub fn lb_external_interface_dir<P: AsRef<Path>>(root: P) -> PathBuf {
+    root.as_ref().join("libs/lb_external_interface")
 }
 
 pub fn local_env_path<P: AsRef<Path>>(root: P) -> PathBuf {
